requirements.txt
asgiref==3.5.2
Django==4.1
django-cors-headers==3.13.0
django-extensions==3.2.0
django-sslserver==0.22
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.4
six==1.16.0
sqlparse==0.4.3
